ovs, tests: split OVS tests from linux bridge tests
This split removes the need for an OVS helpers file.
