+++ internal/hostnetwork/vni_test.go
@@ -443,169 +443,6 @@ var _ = Describe("L2 VNI configuration", func() {
 	)
 })
 
-var _ = Describe("L2 VNI configuration with OVS bridges", func() {
-	var testNS netns.NsHandle
-
-	BeforeEach(func() {
-		cleanTest(testNSName)
-		testNS = createTestNS(testNSName)
-		setupLoopback(testNS)
-	})
-
-	AfterEach(func() {
-		cleanTest(testNSName)
-	})
-
-	It("should work with a single L2VNI using auto-created OVS bridge", func() {
-		params := L2VNIParams{
-			VNIParams: VNIParams{
-				VRF: "testred", TargetNS: testNSName,
-				VTEPIP: "192.170.0.9/32", VNI: 100, VXLanPort: 4789,
-			},
-			HostMaster: &HostMaster{Type: OVSBridgeLinkType, AutoCreate: true},
-		}
-
-		err := SetupL2VNI(context.Background(), params)
-		Expect(err).NotTo(HaveOccurred())
-
-		Eventually(func(g Gomega) {
-			validateL2HostLeg(g, params)
-			_ = inNamespace(testNS, func() error {
-				validateL2VNI(g, params)
-				return nil
-			})
-		}, 30*time.Second, 1*time.Second).Should(Succeed())
-
-		By("removing the VNI")
-		err = RemoveNonConfiguredVNIs(testNSName, []VNIParams{})
-		Expect(err).NotTo(HaveOccurred())
-
-		By("checking the VNI and OVS bridge are removed")
-		vethNames := vethNamesFromVNI(params.VNI)
-		Eventually(func(g Gomega) {
-			checkLinkdeleted(g, vethNames.HostSide)
-			checkOVSHostBridgeDeleted(g, params)
-			_ = inNamespace(testNS, func() error {
-				validateVNIIsNotConfigured(g, params.VNIParams)
-				return nil
-			})
-		}, 30*time.Second, 1*time.Second).Should(Succeed())
-	})
-
-	It("should work with a single L2VNI using pre-existing named OVS bridge", func() {
-		const bridgeName = "test-ovs-br"
-		Expect(createOVSBridge(bridgeName)).To(Succeed(), "must pre-provision an OVS bridge")
-
-		params := L2VNIParams{
-			VNIParams: VNIParams{
-				VRF: "testred", TargetNS: testNSName,
-				VTEPIP: "192.170.0.9/32", VNI: 100, VXLanPort: 4789,
-			},
-			HostMaster: &HostMaster{Type: OVSBridgeLinkType, Name: bridgeName},
-		}
-
-		err := SetupL2VNI(context.Background(), params)
-		Expect(err).NotTo(HaveOccurred())
-
-		Eventually(func(g Gomega) {
-			validateL2HostLeg(g, params)
-			checkOVSBridgeExists(g, bridgeName)
-			checkVethAttachedToOVSBridge(g, bridgeName, vethNamesFromVNI(params.VNI).HostSide)
-		}, 30*time.Second, 1*time.Second).Should(Succeed())
-
-		By("removing the VNI")
-		err = RemoveNonConfiguredVNIs(testNSName, []VNIParams{})
-		Expect(err).NotTo(HaveOccurred())
-
-		By("checking the bridge persists (user-managed)")
-		Eventually(func(g Gomega) {
-			checkOVSBridgeExists(g, bridgeName) // Bridge should still exist
-		}, 30*time.Second, 1*time.Second).Should(Succeed())
-	})
-
-	It("should work with multiple L2VNIs with different auto-created OVS bridges + cleanup", func() {
-		params1 := L2VNIParams{
-			VNIParams: VNIParams{
-				VRF: "testred", TargetNS: testNSName,
-				VTEPIP: "192.170.0.9/32", VNI: 100, VXLanPort: 4789,
-			},
-			HostMaster: &HostMaster{Type: OVSBridgeLinkType, AutoCreate: true},
-		}
-
-		params2 := L2VNIParams{
-			VNIParams: VNIParams{
-				VRF: "testgreen", TargetNS: testNSName,
-				VTEPIP: "192.170.0.9/32", VNI: 101, VXLanPort: 4789,
-			},
-			HostMaster: &HostMaster{Type: OVSBridgeLinkType, AutoCreate: true},
-		}
-
-		err := SetupL2VNI(context.Background(), params1)
-		Expect(err).NotTo(HaveOccurred())
-		err = SetupL2VNI(context.Background(), params2)
-		Expect(err).NotTo(HaveOccurred())
-
-		Eventually(func(g Gomega) {
-			validateL2HostLeg(g, params1)
-			validateL2HostLeg(g, params2)
-		}, 30*time.Second, 1*time.Second).Should(Succeed())
-
-		By("removing VNI 100, keeping VNI 101")
-		err = RemoveNonConfiguredVNIs(testNSName, []VNIParams{params2.VNIParams})
-		Expect(err).NotTo(HaveOccurred())
-
-		By("checking VNI 100 removed, VNI 101 persists")
-		Eventually(func(g Gomega) {
-			checkOVSHostBridgeDeleted(g, params1)
-			checkOVSBridgeExists(g, hostBridgeName(params2.VNI))
-		}, 30*time.Second, 1*time.Second).Should(Succeed())
-	})
-
-	It("should be idempotent with OVS bridges", func() {
-		params := L2VNIParams{
-			VNIParams: VNIParams{
-				VRF: "testred", TargetNS: testNSName,
-				VTEPIP: "192.170.0.9/32", VNI: 100, VXLanPort: 4789,
-			},
-			HostMaster: &HostMaster{Type: OVSBridgeLinkType, AutoCreate: true},
-		}
-
-		err := SetupL2VNI(context.Background(), params)
-		Expect(err).NotTo(HaveOccurred())
-
-		By("calling SetupL2VNI a second time")
-		err = SetupL2VNI(context.Background(), params)
-		Expect(err).NotTo(HaveOccurred(), "second SetupL2VNI should be idempotent")
-
-		Eventually(func(g Gomega) {
-			validateL2HostLeg(g, params)
-		}, 30*time.Second, 1*time.Second).Should(Succeed())
-	})
-
-	It("should configure L2 gateway IP with OVS bridge", func() {
-		gwIP := "10.10.100.1/24"
-		params := L2VNIParams{
-			VNIParams: VNIParams{
-				VRF: "testred", TargetNS: testNSName,
-				VTEPIP: "192.170.0.9/32", VNI: 100, VXLanPort: 4789,
-			},
-			L2GatewayIPs: []string{gwIP},
-			HostMaster:   &HostMaster{Type: OVSBridgeLinkType, AutoCreate: true},
-		}
-
-		err := SetupL2VNI(context.Background(), params)
-		Expect(err).NotTo(HaveOccurred())
-
-		Eventually(func(g Gomega) {
-			validateL2HostLeg(g, params)
-			_ = inNamespace(testNS, func() error {
-				validateL2VNI(g, params)
-				return nil
-			})
-		}, 30*time.Second, 1*time.Second).Should(Succeed())
-	})
-})
-
 func validateL3HostLeg(g Gomega, params L3VNIParams) {
 	vethNames := vethNamesFromVNI(params.VNI)
 	hostLegLink, err := netlink.LinkByName(vethNames.HostSide)
@@ -626,15 +463,6 @@ func validateL3HostLeg(g Gomega, params L3VNIParams) {
 		g.Expect(err).NotTo(HaveOccurred())
 		g.Expect(hasIP).To(BeTrue(), "host leg does not have IPv6", params.HostVeth.HostIPv6)
 	}
-}
-
-func checkOVSHostBridgeDeleted(g Gomega, params L2VNIParams) {
-	g.Expect(params.HostMaster).ToNot(BeNil())
-	g.Expect(params.HostMaster.Type).To(Equal(OVSBridgeLinkType))
-	g.Expect(params.HostMaster.AutoCreate).To(BeTrue())
-
-	hostBridge := hostBridgeName(params.VNI)
-	checkOVSBridgeDeleted(g, hostBridge)
 }
 
 func validateL2HostLeg(g Gomega, params L2VNIParams) {
@@ -850,25 +678,4 @@ func validateBridgeMacAddress(g Gomega, bridge netlink.Link, vni int) {
 	actualMac := bridge.Attrs().HardwareAddr
 	g.Expect(actualMac).NotTo(BeNil(), "bridge should have a MAC address")
 	g.Expect(actualMac).To(Equal(expectedMac), "bridge MAC address should be %v for VNI %d", expectedMac, vni)
-}
-
-// checkOVSBridgeExists validates that an OVS bridge exists
-func checkOVSBridgeExists(g Gomega, bridgeName string) {
-	bridge, err := getOVSBridge(bridgeName)
-	g.Expect(err).NotTo(HaveOccurred(), "failed to get OVS bridge %q", bridgeName)
-	g.Expect(bridge).NotTo(BeNil())
-	g.Expect(bridge.Name).To(Equal(bridgeName))
-}
-
-// checkOVSBridgeDeleted validates that an OVS bridge does not exist
-func checkOVSBridgeDeleted(g Gomega, bridgeName string) {
-	_, err := getOVSBridge(bridgeName)
-	g.Expect(err).To(HaveOccurred(), "OVS bridge %q should not exist", bridgeName)
-}
-
-// checkVethAttachedToOVSBridge validates that a veth is attached to an OVS bridge
-func checkVethAttachedToOVSBridge(g Gomega, bridgeName, vethName string) {
-	hasPort, err := ovsBridgeHasPort(bridgeName, vethName)
-	g.Expect(err).NotTo(HaveOccurred())
-	g.Expect(hasPort).To(BeTrue(), "veth %s should be attached to OVS bridge %s", vethName, bridgeName)
 }
